# src/swagger.json
{
    "openapi": "3.0.1",
    "info": {
        "title": "REST API minta backend",
        "version": "1.0.0",
        "description": "<img alt='DB_diagram' height='300px' src='https://nitslaszlo.github.io/static/session/db_diagram.jpg' />"
    },
    "servers": [
        {
            "url": "http://localhost:5000"
        },
        {
            "url": "https://jedliksession.cyclic.app"
        }
    ],
    "paths": {
...
        "/locations": {
            "get": {
                "tags": ["locations"],
                "summary": "Összes helyszin lekérdezése.",
                "description": "Az összes helyszin adatát tudjuk lekérdezni.",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/location"
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorSchema"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": ["locations"],
                "summary": "Új helyszin mentése.",
                "description": "Új helyszin mentése a 'post' táblában",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "$ref": "#/components/schemas/location"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK (Az új, elmentett posttal tér vissza.)",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "$ref": "#/components/schemas/location"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorSchema"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "security": [
        {
            "name": []
        }
    ],
    "components": {
        "schemas": {
            "location": {
                "type": "object",
                "required": ["title"],
                "properties": {
                    "_id": {
                        "type": "string",
                        "description": "PK - A post azonosítója (ObjectID).",
                        "example": "61b5f28df39e4edcf5b8a3e7"
                    },
                    "user_id": {
                        "type": "string",
                        "description": "FK - A post szerzője (ObjectID).",
                        "example": "61b5e9c0f39e4edcf5b8a3b9"
                    },
                    "content": {
                        "type": "string",
                        "description": "A post tartalma.",
                        "example": "Frankie Dunn pályája során számos fantasztikus bunyóst edzett és menedzselt. Legfőbb tanítása, mely vezérli az életét is, hogy a legfontosabb az, hogy mindig védd magad. Mióta elidegenedett a lányától, Frankie jó ideje senkit sem engedett közel magához. Egyetlen barátja Hepaj, az egykori bokszoló, aki rendben tartja Frankie edzőtermét, és tudja, hogy a mogorva külső olyan embert takar, aki az elmúlt 23 évben szinte minden nap misére járt, keresve a megbocsátást, amely valahogy mégsem adatik meg neki. Aztán egy nap Maggie Fitzgerald besétál az edzőterembe. Maggie-nek sosem jutott sok semmiből, de van benne valami, ami csak keveseknek adatik meg: tudja, mit akar, és bármit hajlandó megtenni érte. Az élete állandó küzdelem, amit nyers tehetségével, kizökkenthetetlen koncentrálásával és elképesztő akaraterejével képes túlélni."
                    },
                    "title": {
                        "type": "string",
                        "maxLength": 50,
                        "description": "A post címe címe.",
                        "example": "Millió dolláros bébi 2004"
                    }
                }
            },
...
            "ErrorSchema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string",
                  "example": "error.message value"
                },
                "status": {
                  "type": "integer",
                  "example": 400
                }
              }
            }
        }
    }
}